+++ crate/server/src/middlewares/jwt_auth.rs
@@ -1,5 +1,6 @@
 use std::{
     pin::Pin,
+    sync::Arc,
     task::{Context, Poll},
 };
 
@@ -16,16 +17,16 @@ use futures::{
 };
 use tracing::{debug, error, trace};
 
-use crate::middlewares::jwt::{decode_jwt, JwtConfig};
+use crate::middlewares::jwt::JwtConfig;
 
 #[derive(Clone)]
 pub struct JwtAuth {
-    jwt_config: Option<JwtConfig>,
+    jwt_config: Option<Arc<JwtConfig>>,
 }
 
 impl JwtAuth {
     #[must_use]
-    pub fn new(jwt_config: Option<JwtConfig>) -> Self {
+    pub fn new(jwt_config: Option<Arc<JwtConfig>>) -> Self {
         Self { jwt_config }
     }
 }
@@ -52,7 +53,7 @@ where
 
 pub struct JwtAuthMiddleware<S> {
     service: S,
-    jwt_config: Option<JwtConfig>,
+    jwt_config: Option<Arc<JwtConfig>>,
 }
 
 impl<S, B> Service<ServiceRequest> for JwtAuthMiddleware<S>
@@ -71,20 +72,17 @@ where
 
     fn call(&self, req: ServiceRequest) -> Self::Future {
         // get the JWT config
-        let jwt_config = match &self.jwt_config {
-            Some(c) => c,
-            None => {
-                return Box::pin(async move {
-                    error!(
-                        "{:?} {} 401 unauthorized: JWT not properly configured on KMS server ",
-                        req.method(),
-                        req.path(),
-                    );
-                    Ok(req
-                        .into_response(HttpResponse::Unauthorized().finish())
-                        .map_into_right_body())
-                })
-            }
+        let Some(jwt_config) = &self.jwt_config else {
+            return Box::pin(async move {
+                error!(
+                    "{:?} {} 401 unauthorized: JWT not properly configured on KMS server",
+                    req.method(),
+                    req.path(),
+                );
+                Ok(req
+                    .into_response(HttpResponse::Unauthorized().finish())
+                    .map_into_right_body())
+            })
         };
 
         trace!("JWT Authentication...");
@@ -97,10 +95,12 @@ where
                     .and_then(|h| h.to_str().ok().map(std::string::ToString::to_string))
             })
             .unwrap_or_default();
+        trace!("Checking JWT identity: {identity:?}");
 
         // decode the JWT
-        trace!("Checking JWT");
-        let private_claim = decode_jwt(jwt_config, &identity).map(|claim| claim.email);
+        let private_claim = jwt_config
+            .decode_bearer_header(&identity)
+            .map(|claim| claim.email);
         match private_claim {
             Err(e) => Box::pin(async move {
                 error!(