+++ crate/server/src/middlewares/mod.rs
@@ -4,4 +4,4 @@ pub use jwt_auth::{JwtAuth, JwtAuthClaim};
 pub mod ssl_auth;
 
 mod jwt;
-pub use jwt::JwtConfig;
+pub use jwt::{JwtConfig, JwtTokenHeaders, UserClaim};